Fix a couple of 64bit related issues. Update to latest orange.

+++ orbit/orb/Exceptions.d
@@ -13,7 +13,7 @@ class OrbException : OrbitException
 {
 	template Constructor ()
 	{
-		this (string message, string file, long line)
+		this (string message, string file, size_t line)
 		{
 			super(message, file, line);
 		}
